Extract showPerformanceWarning
